join yelp and hygiene data
yh_join.js is used to join the hygiene data with yelp data, you can run it in mongo shell by type load("//the address of yh_join.js"). Because in the hygiene dataset, the name are all capital letters, but yelp has capital letters and small letters, so I change all the name in yelp to capitals, then use lookup to join them. I made a test, the time of transfer small letters to capital letters is not long, maybe about 1 minute.