new data cleaning and mapreduce
This weisheng directory contains all the files used in data cleaning and map reduce. 
The weisheng,sh indicate the execution sequence of all the files. 
And in data directory, the yelpDataMain.json is the dataset I got from yelp API.
In the data/result directory, the result.json is the result which add the average score to each yelp document, so you guys can import the json file directly to the mongodb and don't have to execute every files.